Make amap base prepared request param to base
