+++ tox.ini
@@ -13,12 +13,12 @@ deps =
 
 [testenv:py27]
 commands =
-    pytest --cov-report term-missing --cov={envsitepackagesdir}/thrall -s --pyargs thrall tests
+    pytest --cov-report term-missing --cov={envsitepackagesdir}/thrall --pyargs thrall tests
 
 [testenv:py36]
 commands =
-    pytest --cov-report term-missing --cov={envsitepackagesdir}/thrall -s --pyargs thrall tests
+    pytest --cov-report term-missing --cov={envsitepackagesdir}/thrall --pyargs thrall tests
 
 [testenv:pypy]
 commands =
-    pytest --cov-report term-missing --cov={envsitepackagesdir}/thrall -s --pyargs thrall tests
+    pytest --cov-report term-missing --cov={envsitepackagesdir}/thrall --pyargs thrall tests
